pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "reval"
version = "0.1.0"
description = "Robust Evaluation of Values and Alignment in LLMs - A benchmark for measuring political bias"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "REVAL Team" }
]
keywords = ["llm", "benchmark", "bias", "evaluation", "alignment"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0",
    "boto3>=1.34",
    "aioboto3>=12.0",
    "jsonlines>=4.0",
    "jsonschema>=4.0",
    "pyyaml>=6.0",
    "numpy>=1.24",
    "rich>=13.0",
    "typer>=0.9",
    "python-dotenv>=1.0",
    # Phase 3: provider unification. Async SDKs for non-Bedrock surfaces.
    "anthropic>=0.40",
    "openai>=1.50",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "black>=24.0",
    "mypy>=1.0",
    "types-PyYAML>=6.0",
    "pre-commit>=3.5",
]

[project.scripts]
reval = "reval.cli:app"

[project.urls]
Homepage = "https://github.com/reval-benchmark/reval"
Documentation = "https://github.com/reval-benchmark/reval#readme"
Repository = "https://github.com/reval-benchmark/reval"

[tool.hatch.build.targets.wheel]
packages = ["src/reval"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # line length handled by black; E501 just nags about long strings

[tool.ruff.lint.isort]
known-first-party = ["reval"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests", "evaluations"]
python_files = ["test_*.py", "eval_*.py"]
markers = [
    "eval: evaluation tests — run with: pytest -m eval (requires AWS credentials)",
]

[tool.mypy]
python_version = "3.10"
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
# Non-strict baseline — existing code does not pass --strict.
# Tightening to strict is a separate cleanup outside this plan's scope.
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
